Facial_recognition-main/Facial recognition using opencv and python libraries for smart voting/requirements.txt
opencv-python
scikit-learn
pywin32
pandas
openpyxl
pyqt5
bcrypt
pillow
flask
flask-wtf
flask-bootstrap
flask-login